[FLINK-10565] [tests] Refactor SchedulerTestBase
