+++ flink-runtime/src/test/java/org/apache/flink/runtime/executiongraph/ExecutionGraphCoLocationRestartTest.java
@@ -32,8 +32,6 @@ import org.apache.flink.runtime.jobmaster.slotpool.SlotProvider;
 import org.apache.flink.util.FlinkException;
 
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
 import java.util.function.Predicate;
 
@@ -47,14 +45,9 @@ import static org.junit.Assert.assertThat;
  * Additional {@link ExecutionGraph} restart tests {@link ExecutionGraphRestartTest} which
  * require the usage of a {@link SlotProvider}.
  */
-@RunWith(Parameterized.class)
 public class ExecutionGraphCoLocationRestartTest extends SchedulerTestBase {
 
 	private static final int NUM_TASKS = 31;
-
-	public ExecutionGraphCoLocationRestartTest(SchedulerType schedulerType) {
-		super(schedulerType);
-	}
 
 	@Test
 	public void testConstraintsAfterRestart() throws Exception {
@@ -79,10 +72,10 @@ public class ExecutionGraphCoLocationRestartTest extends SchedulerTestBase {
 			groupVertex,
 			groupVertex2);
 
-		if (schedulerType == SchedulerType.SLOT_POOL) {
-			// enable the queued scheduling for the slot pool
-			eg.setQueuedSchedulingAllowed(true);
-		}
+
+		// enable the queued scheduling for the slot pool
+		eg.setQueuedSchedulingAllowed(true);
+
 
 		assertEquals(JobStatus.CREATED, eg.getState());
 
@@ -122,7 +115,7 @@ public class ExecutionGraphCoLocationRestartTest extends SchedulerTestBase {
 
 		ExecutionJobVertex[] tasks = eg.getAllVertices().values().toArray(new ExecutionJobVertex[2]);
 
-		for(int i = 0; i < NUM_TASKS; i++){
+		for (int i = 0; i < NUM_TASKS; i++) {
 			CoLocationConstraint constr1 = tasks[0].getTaskVertices()[i].getLocationConstraint();
 			CoLocationConstraint constr2 = tasks[1].getTaskVertices()[i].getLocationConstraint();
 			assertThat(constr1.isAssigned(), is(true));
