+++ flink-runtime/src/test/java/org/apache/flink/runtime/jobmanager/scheduler/ScheduleWithCoLocationHintTest.java
@@ -27,8 +27,6 @@ import org.apache.flink.runtime.taskmanager.TaskManagerLocation;
 import org.apache.flink.runtime.testingUtils.TestingUtils;
 
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
 import java.util.Collections;
 import java.util.concurrent.ExecutionException;
@@ -40,12 +38,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
-@RunWith(Parameterized.class)
 public class ScheduleWithCoLocationHintTest extends SchedulerTestBase {
-
-	public ScheduleWithCoLocationHintTest(SchedulerType schedulerType) {
-		super(schedulerType);
-	}
 
 	@Test
 	public void scheduleAllSharedAndCoLocated() throws Exception {
